+++ tests/testthat/test-lists.R
@@ -6,7 +6,7 @@ test_that("concatenation works", {
                    div = c(country = "United States", state = "Nebraska"))
 
   # Create a model
-  object1 <- new("PersephoneQuasiBin",
+  object1 <- new("PersephoneBin",
                  region = region,
                  crop = "Corn",
                  data = progress_ne$Corn,
@@ -24,7 +24,7 @@ test_that("concatenation works", {
 
   # Check classes
   expect_s3_class(object, "PersephoneModelList")
-  expect_s4_class(object[[1]], "PersephoneQuasiBin")
+  expect_s4_class(object[[1]], "PersephoneBin")
   expect_s4_class(object[[2]], "PersephoneCumLink")
 
 })