+++ tests/testthat/test-summary.R
@@ -6,7 +6,7 @@ test_that("summary works", {
                    div = c(country = "United States", state = "Nebraska"))
 
   # Create a model
-  object1 <- new("PersephoneQuasiBin",
+  object1 <- new("PersephoneBin",
                  region = region,
                  crop = "Corn",
                  data = progress_ne$Corn,